requirements.txt
attrs==22.1.0
boto3==1.26.7
botocore==1.29.7
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
exceptiongroup==1.0.1
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
s3transfer==0.6.0
secret==0.8
six==1.16.0
SQLAlchemy==1.4.43
tabulate==0.9.0
tomli==2.0.1
urllib3==1.26.12
uuid==1.30
Werkzeug==2.2.2
